kokkos: refactor duplicated test helpers
